Fixed lurking bug in use of the symbol state engine.
When creating a :lexeme (or :discard) the literal's symbol has to be
recorded with a l0-usage at that point. Forgetting that is no problem
if the symbol is for an actual literal, its state is literal either
way. However with the deconstruction of literals into simpler
pieces such symbols may stand for priority rules as well. For them,
with the l0-usage forgotten they look like toplevel L0 symbols,
i.e. lexemes, except that the engine also knows that they were not
used in G1, thus reports the situation as error.