Implemented basic container validation.
As part of that added some plumbing so that the objects internal to
the container have a reference to their collection (container, or
grammar, depending on context). Needed as some validations have to
reach from the location perf,orming them to other parts of the system
to get associated data.

Extended the testsuites to invoke the validator (result of ctrace ->
gcstate, ser/des, reductrion results). Pleased that all of them are good.

Related things, to do: Checks for inacessible symbols, non-generating
symbols (not reaching terminals/atoms), removal of same.

Further to do, rewriting of precedenced priority rules to
non-precedenced form.