Replaced fixed slif meta grammar under tests/grammars/ with reference to new slif meta grammar under bootstrap.
The new grammar removes the "bless" adverb.  It further uses a redone
syntax for strings and char classes which encodes more of Tcl's
constraints into the syntax.

Modified the semantics to handle the possibility of <single quoted
name> using Tcl backslash-escapes, as now allowed by the new meta
grammar. While this has not arrived in the grammar used by
"s_parser.tcl" yet it may be.

Cleaned the bootstrap/ directory up a bit.
(Still have to check and compare the various old meta grammars)