Moved progress reporting into a mixin (engine_debug.tcl). Added code for the reporting of marpa parse trees (valuation instruction list).
Added stream reporting to gate, lexer, and parser.
Added progress and forest reporting to lexer and parser.

Added capture of complete rule information in the base engine, for the
progress report (Semantics might need this as well). Updated the
semantics core to match the changes in the engine.

Reworked lexer internals with regard to LATM, LTM lexemes, and
discards. All of them get an ACS symbol. This means there is an 1:1
mapping between lexeme and discard rules, and the ACS symbols. The
difference between LATM and LTM + discard is that the latter group is
always acceptable/activated. Note, this means that the first
instruction of any parse tree is the token for the ACS symbol, which
gives us the matched rule. We do not need the entire parse tree, nor
do we have to actually evaluate it. This will allow us to replace the
GetSymbol and GetString semantics with simpler code.