Removed GetString, GetSymbol sem core instances from the lexer. Replaced by code using direct access to parse tree and engine state information to generate the semantic value. Updated test suite.
Changes in the output for rule. The old output was a bug in the
GetString setup, reporting the internal `@START ~ foo` rule, instead
of whichever rule for `foo` was matched.

General changes in the lexer output, it now reports separate semantic
values for all matched symbols. That was also a bug in the previous
setup. It worked only because the runtime was geared towards
(start,llength,value), which is identical across symbols. With the
proper reporting for other keys, like 'rule', this is not the case
anymore. The parser class was updated to match this API change.

This change further reflects in the expected reults of the tests for
lhs, name, rule, and symbol, testsuite is updated.

Some optimization is still done, i.e. it is made sure that identical
textual semantic values are stored under the same id as well,
i.e. only once.