Added formatter producing a SLIF grammar from a grammar container.
Note, this is not fully round-trip at the moment (The special @LEX
symbols can not be read back, violating identifier syntax).

It is also sub-optimal with regard to LATM flags, g1 actions,
etc. These are shown as attributes of each rule instead of making use
of defaults to reduce duplication.

It should be good enough however to serve as debugging aid.