Changed gate to lexer flush signaling from in-band `(byte) -1` to a separate function. This removes any possibility that a `(byte) -1` from actual input causes a bogus flush.
Added debug function allowing INBOUND to properly print a batch of
input bytes.

Fixed a crash of the RT-C where the loop searching for the end of the
lexeme tried to pop a byte from the empty lexeme, triggering an
underflow assert. This may happen when `lexer_complete` is called for
an empty-valued lexeme.  I.e. when the GATE rejects the first byte
after the end of a lexeme as invalid and signals a flush before any
byte was entered into the lexer at all.

Note that this does not necessarily indicate a mismatch. The current
set of acceptable lexemes may contain some which allow an empty
value. We have to keep recognizing them. And after that the new
context may have caused the invalid byte to be valid.

So we only skip the attempt of making an empty value even emptier.
The deeper issue is that for LATM-mode symbols the earley-set id does
not match the length of the lexeme due to the zero-width ACS guards in
front; causing an additional round through the loop before it can
declare mismatch.

The concrete example which triggered the issue are the `string` and
`lstring` symbols in the JSON grammar, for input `[""]`.