Continued rework of the unicode layer. See first commit in the branch for the plan.
Removed full vs bmp from the `unidata` tool and generated tables. The
tool now always generates tables covering the full unicode range.

Updated some tests, but not all. The known test failures are in the
various generators and the middleware, due to the CC differences
coming from full coverage. Fixing it now does not make sense, because
we will have to clean it up again after the introduction of MUTF-8 and
CESU-8 support into the middle layers. We clean up these up after that
is all done.