Created a Marpa parser for the string and CC lexemes. This parser handles all the various forms for character escapes. Plus a semantics backend which generates the internal literal representation directly from the AST.
The above replaced the entire existing literal processor (parse,
decode, unescape, type, tags). This was needed because Tcl (especially
`subst` as the core of the old `unescape`) was/is not able to handle
the full set of unicodepoints (at this time).

Updated the bootstrap slif parser to handle the extended escapes too
(\u hex x5/x6, \U hex x8).

Updated tests, although not all.

With this commit the entire input side is now able to handle the full
set of unicode, with suitable escape sequences for characters outside
of the BMP as well.