Reworked the internals of the unicode layer. Strong split into C files and the C/Tcl glue.
Exposed the codepoint validation as custom critcl type, and added use
of it to 2utf.

Extended 2utf and 2asbr with optional flag argument to control the
coding (mutf, cesu, tcl is both). Due to the new argument being
optional (and last) the existing call sites will not fail, and operate
in full mode.

Updated the layer's test suite demonstrating the mutf and cesu
coding. Added more tests for invalid arguments.