Tool tweak. For each named class FOO generate additional named classes `FOO:bmp` and `FOO:high` which are limited to the parts of FOO inside and outside the BMP.
Note: Empty classes are not generated at all. Example: `adlam:bmp`
does not exist because it would be empty. In turn `adlam::high` ==
`adlam`.

Note 2: The `:` in the names of the new classes prevents use from
SLIF. This may change in the future, as it might be sensible to
provide languages access to the limited classes.