Updated the json tests. The only tests directly affected by the merge where a number of expected-negatives where the error message did not match anymore, due to changed symbol ids coming out of the generator.
Furthermore, the `process-file` changes reduced the set of known bugs,
eliminating all expected-negatives from it (process-file sees the same
data as process now,, making the error messages the same for the two
methods). We still have issues with various expected-positive which
still error. These are suspected to require changes to the json grammar.