Added test triggering rtC code path where a new lexer starts out with the parser exhausted, i.e. nothing acceptable.
Fixed missing closing of the earleme for that case, and missing
handling of the `lexer exhausted` error from libmarpa. The missing
closing operation also caused miscommunication between lexer and gate,
ultimately crashing the latter with a symbol id outside of the byte
range.

The gate change is only a tweak to get better tracing, i.e. print the
acceptables before working with them, not after.