Fixed a design issue.
Surrogate handling.

The low-level unicode layer used a guard to prevent the addition of
the surrogate codepoints when creating an ASBR from a CC. Thus for any
CC including one or more surrogates the resulting ASBR actually
represented the CC minus surrogates.

The thinking was sort of that we are working with characters at the
high-level, and while we have an 1:1 mapping to codepoints for most,
this is not true for the surrogates, each of which is only half of a
character.

And the reducer targeting the C runtime based on Tcl ensures that
characters in SMP are properly converted into surrogate pairs.

Working on the JSON parser now has driven home that there are
situations where we want to accept standalone surrogates, at the
highlevel, and also that the low-level removal was a bad idea as
well. The latter because a negated char class handled by Tcl does
accept the standalone surrogate code points, whereas the ASBR for C is
mangled to reject them. The result is a very unwanted difference in
the behaviour of what should be equivalent parsers.

So, lots of writing for a very small change, code-wise, the removal of
the guard mentioned in the 2nd paragraph. And undoing the removal of
the surrogates as a named character class.


Further, brought the bugfix from commit [546018b243] into the
`unicode_ops.tcl` used by the table generator. Same issue, and
forgotten when the initial fix was made and committed.


Updated tests to match results due to rule renumeration and CC changes.