The known bug was due to a subtle difference in the two parsers.
The Tcl-based parser accepted standalone surrogates, the C-based one
did not. This came down to a guard condition in the ASBR compiler,
which excluded any surrogates found in the input CC from the result.

A design bug, not an implementation bug.

Removing the guard fixes the issues with the json parser.

The first attempt at the fix, adding the standalone surrogates
explicitly to the grammar run into the same guard, albeit in a
different manner. The explicit range became an empty literal during
reduction, got removed, and then the still-existing reference to it
caused the generator backend to throw an error.

Regenerated the C-based parser, and updated all test results to match
the changes in the rule numeration.