inbound, gate, lexer:
Added a back-link from gate to inbound, analogous to the gate/lexer
and lexer/parser connections. Dropped history management from the gate
and changed its redo method to simply rewind the input instead of
re-entering the tail end of the history. This is the first use of the
new cursor movement methods added to inbound.

Under the old regime using foreach, and steadily marching forward in
the input all re-processing was done by remembering and recursively
re-entering characters as needed, with additional loops at the various
stages of the pipeline (`gate`).

With the while-based cursor on the other hand we have only one
(nested) processing loop (`inbound`), and all movement is handled by
it. We cannot recurse, and we can't have a loop in `gate`.

The nested loop in `inbound` is required to handle the case where we
reach eof and the later stages then tell us `not yet`, i.e. bounce us
back. The inner loop is the main processor, and the outer loop restarts
it until eof actually succeeds.

In `gate` the flush-signalling changed. In 'lexer' eof signalling to the
`parser` is prevented when it bounces the input away from eof.