Moved the main parts of `test/support/gen.tcl` into `bin/i-gen`. This new internal generator application uses the public `marpa-gen` as the underlying workhorse. The support code now invokes the internal generator instead of doing everything itself.
With the operation confined to a child process the test process'
amount of memory taken is limited to the generated package instead of
also keeping all the overhead of generating

Note, this does not help with the test suites based on lexer-core as
that suite still loads/sources ten different lexer packages into the
same process. Handling this requires more work, i.e. pushing the
individual tests into their own child processes.