Moved the main parts of `test/support/gen.tcl` into `bin/i-gen`. This new internal generator application uses the public `marpa-gen` as the underlying workhorse. The support code now invokes the internal generator instead of doing everything itself.
With the operation confined to a child process the amount of memory
taken by the test process' is limited to the generated package itself
instead of all the packages needed to perform the generation.

Further modified lexer-core to place the sets of tests for each
variant grammar/lexer-action into a child process as well. This
ensures that the memory needed by each variant grammar is limited to
that child process, instead of accumulating in the controlling test
process.

__Attention__: This change requires a Kettle with support for
`kt::sub`, added to Kettle with commit [ef384673c5] (2018-04-18
02:28:17).