Fixed [f5e6063aeb] memory smash. Miscounted references to the `self` argument of parse event callbacks. Updated tests with all the generator changes.
rtc-runtime tests: L0 parse event cases failing as expected, as event
generation is not done yet. However we are now at the point where we
can start on adding this in. Most of the foundations are now
present. Notably still missing are the parse event descriptor
structures and its linkage to the lexer state.