Continued fill-in for before/after events.
`Inbound`, `gate`, and `lexer` rewritten to match rt-Tcl behaviour
with regard to stepping through the input, rewind, and flush.

__Attention__ `Inbound` location tracking handles only ASCII
correctly, still, not multi-byte UTF. Still mulling over possible
index structures to enable quick movement in the presence of
multi-bytes with low memory overhead for very uniform input.

Changed the `lexer` field `m_sv` to an int stack, the component tracks
ids, with the associated SVs already remembered in the `store`. Forced
changes to the implementation of the facade, and to the API between
`lexer` and `parser` (changed signature and implementation of
`marpatcl_rtc_parser_enter`).

Fixed forgotten difference between discard and other events (ACS
symbols vs G1 symbols) in the C generator code and the id/symbol
conversion done by the facade.