Scratch commit. Compiles. Tests fail.
Begun implementation of initial from/to/limit for `process(-file)`.
Required to test stop events (force event from outside).
Also a feature of the original.

Fixed forgotten use of `location?` in the inbound tests.

Need more inbound tests for all the new methods for cursor movement.

Bug: The ability to force a stop event on the engine from the outside
implies that we have to support events even for grammars which do not
declare any. The templates for rtC-based parser/lexers must change
accordingly. We might be able to optimize them down to `support only
stop`.