Draft of adapter to provide management of multiple (stop) markers to any engine. No tests yet.
Extended runtimes to remember and deliver the last location of the
primary input stream. Moving to that location (`match from`) will
force the engine to abort processing.

As we have only a byte length at the beginning we needed an additional
function for the char index to map byte offsets into char
offsets. Some more in-band signaling to defer the conversion of the
byte length to char length until the information is actually requested
for the first time. This ensures that parsers not doing IO shenanigans
will not have to scan the entire input twice, for the last char, and
then for actual processing. Only parsers which do will have to expend
that time, and only as much as it not yet processed.