Oh! :facepalm:
`inbound_enter` copies just the pointer to the input string into the
engine state. This pointer is (as critcl type `pstring`) just a
pointer into the Tcl_Obj internals (GetStringFromObj).

We are fine as long as we do not perform `inbound_enter_more`, because
nothing does anything with the pointer, but reading from it. Not even
basic events.

But with `inbound_enter_more` we realloc that area. Means that the old
space will be freed when the system did not find enough space to
expand in place. At which point we pulled the rug out of under the
Tcl_Obj* the string came from. Doing anything with that Tcl_Obj* (like
releasing it, example `unset var`) will then smash that memory.

Fixed, `inbound` now tracks if it owns IN.bytes and acts accordingly
when expanding the area in `enter_more`.

TODO: Validate stop locations, must not allow anything beyond the end
      of the allocated memory, system will try to run over into
      uninitialized cells at the moment.

TODO: Alt: Force engine stop on max for total as measure against the
      above.