Scratch
- Added narrative tracing to inbound around the IO loop, track exactly
  where it is and how it moves.

- X1, X2 - Demo of the `-1` difference for the include-iii test
  processing.

It seems the issue is rooted in the fact that I stop the engine before
the marked character is processed. Trying to set the stop marker after
the last character of the stream, to stop after everything has been
processed ... does not work.

The possible fix would be to change the semantics and stop after the
marked character has processed. Then setting the marker on the last
character will stop after it, before trying to move into the bad lands.

Techically this shifts the impossible stop location to before the
first character of the input. OTOH, if we choose to allow an absolute
position -1, that might work (shift the controlling test to the
beginning of the io loop). The inband `no stop`-signal would then have
to be `-2`.