Tentative change to inbound for `stop after a character`.
With controlling condition at a place in the loop where it can `stop
before the first character` also, giving the full range.

Condering to tweak the user-visible api to still mean `stop before
character`, as a shifted form of the above, so that all user-specified
locations are >= 0, never negative.

This work will move to its own branch off `parse-events`, to work it
out completely, including internal docs before continuing on the
(min)doctools.