More grammar fixes based on processing the Tcllib manpages.
Implemented a lexeme barrier (lexer flush) for use in stop event
handlers. Forces the lexeme sequence for the included file or similar
to coincide with the end of that secondary input, and everything going
up to the border flushed, so that it will be crossed without any
rewind after the last lexeme. Prevents rewinding in the outer input
stream after returning from the include, causing very bogus
lexing. Note, this only a lexer flush, without a parser eof.

Updated doctools-base to use the new barrier in its include
management.

Grammar changes/fixes:
- Dropped the special forms fully from the G1.
- Configured the special forms for LTM, i.e. to always match.
- Allow use of 'enum' as a list type, vs. just the official 'enumerated'.
- Allow opening brace in example text, and quoted strings.
- Fixed `subsection` syntax, may have two arguments.

This fixed an issue with seeing include or vset in places not expected
by the G1 from before the change. Now they can appear everywhere
between lexemes.

Fixed rt-Tcl's handling of LTM (= always-on) lexemes.
Bad id conversion in a few places.

Statistics:
- 355 pass   82%
-  77 fail   18%
- 432 total 100%