Fixed miscounting of multi-byte characters in rtC.
Reworked `inbound_step()`.

Changed to test a character for correctness on the first byte, then
arrange to (not) step the character location on the trailer bytes
following, if any. Always steps the character location on the first
byte. Updates the byte/char index the same time.

Further reworked and fixed the interface from `gate` to `inbound` for
rewinding after a rejection (straight end, or when the matched lexeme
is actually shorter than the number of bytes consumed before
stopping). The used API (`moveby`) was character based, which caused
us to rewind to far, namely to the previous character. No trouble when
it was a single-byte character. Bad for a multi-byte however. This
could then trip the lexer if the first byte of the previous character
is not acceptable at the given point.